+++ src/main/clojure/com/github/clojure_lsp/intellij/notification.clj
@@ -2,7 +2,7 @@
   (:require
    [com.github.clojure-lsp.intellij.client :as lsp-client]
    [com.github.clojure-lsp.intellij.db :as db]
-   [com.github.clojure-lsp.intellij.logger :as logger]
+   [com.github.ericdallo.clj4intellij.logger :as logger]
    [com.github.clojure-lsp.intellij.tasks :as tasks]
    [com.github.ericdallo.clj4intellij.app-manager :as app-manager]
    [seesaw.core :as see])
@@ -37,7 +37,7 @@
                   :to-string :title))}))
 
 (defmethod lsp-client/progress :default [_ progress]
-  (logger/warn "Unknown progress token %s" progress))
+  (logger/warn "Unknown progress token" progress))
 
 (defmethod lsp-client/progress "lsp-startup" [{:keys [progress-indicator]} {{:keys [title message percentage]} :value}]
   (let [msg (str "LSP: " (or title message))]