+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Use nrepl and logger from clj4intellij
+
 ## 0.12.1
 
 - Fix exception on startup related to status bar.